+++ clients/ctl/admin-ui/src/features/system/PrivacyDeclarationAccordion.tsx
